The great refactor - moved most of the logic out of setup into new files - changed module exports into invocable exported functions instead of objects - functions are abstracted out to their specific file; separation of defintions allows for more clear terms specific to that file, but are only called by the exported function rather than being made globally available - some other junk
