+++ _src/scss/_menu-audio.scss
@@ -1,0 +1,15 @@
+#audio-menu-open {
+  position: fixed;
+  right: 0;
+  bottom: 0;
+  margin: 5px;
+  z-index: 999;
+}
+
+#audio-menu {
+  position: fixed;
+  right: 0;
+  bottom: 0;
+  width: 33%;
+  background: rgba(0, 0, 0, 0.65);
+}
