+++ _src/scss/main.scss
@@ -6,5 +6,6 @@ $margin-base: 10px;
 @import '_layout';
 @import '_helpers';
 @import '_editor';
-@import '_menu';
+@import '_menu-editor';
+@import '_menu-audio';
 @import '_elements';
