+++ public/assets/css/main.css
@@ -10,117 +10,9 @@ body {
   overflow-y: hidden;
   padding: 0; }
 
-button {
-  background: none;
-  border: 0;
-  font: inherit;
-  line-height: normal;
-  outline: none;
-  overflow: visible; }
-
-.btn.gourd {
-  background: #d9d9d9;
-  border-bottom: 2px solid transparent;
-  height: 25px;
-  margin: 5px;
-  margin: 0 20px;
-  padding-top: 5px;
-  width: 25px;
-  border-radius: 50%;
-  font-size: 10px; }
-  .btn.gourd:hover {
-    background: #c8c8c8; }
-  .btn.gourd:active {
-    background: #b4b4b4; }
-  .btn.gourd.on {
-    background: #dbdbdb;
-    box-shadow: 0px 0px 5px #ebebeb; }
-
 #page {
   display: flex;
   flex-grow: 1; }
-
-#menu-open {
-  position: fixed;
-  left: 0;
-  bottom: 0;
-  margin: 5px; }
-
-#menu-close {
-  position: fixed;
-  left: 0;
-  bottom: 0;
-  margin: 5px;
-  z-index: 999; }
-
-#editor-controls {
-  display: flex;
-  flex-direction: column;
-  width: 100%; }
-
-.button-drawer {
-  background: rgba(0, 0, 0, 0.6);
-  bottom: -50px;
-  display: flex;
-  flex-direction: row;
-  height: 45px;
-  justify-content: space-around;
-  left: 0;
-  margin: 0 auto;
-  position: fixed;
-  width: 100%;
-  transition-property: all;
-  transition-duration: 0.75s;
-  transition-timing-function: cubic-bezier(0, 1, 0.5, 1); }
-  .button-drawer.open {
-    bottom: 0; }
-
-.button-drawer-style {
-  color: #e7e7e7;
-  font-size: 12px;
-  font-weight: 300;
-  letter-spacing: 2px;
-  text-transform: uppercase;
-  transition-duration: 0.75s;
-  transition-property: all;
-  border-bottom: 2px solid transparent;
-  align-self: center; }
-  .button-drawer-style.on {
-    color: #fff;
-    text-shadow: -1px 1px 3px rgba(208, 208, 208, 0.2), 1px -1px 3px rgba(252, 252, 252, 0.2);
-    border-bottom: 2px solid white; }
-  .button-drawer-style.dim {
-    color: rgba(255, 255, 255, 0.5); }
-
-.asset-drawer {
-  background: rgba(0, 0, 0, 0.6);
-  display: flex;
-  flex-direction: row;
-  justify-content: center;
-  align-items: center;
-  width: 100%;
-  position: fixed;
-  bottom: 45px;
-  height: 40px;
-  padding: 5px;
-  transition-property: all;
-  transition-duration: 0.75s;
-  transition-timing-function: cubic-bezier(0, 1, 0.5, 1); }
-
-.asset-row {
-  display: flex;
-  flex-direction: row; }
-
-.button-header {
-  align-items: center;
-  align-self: center;
-  background: #4b4b4b;
-  color: #bbbbbb;
-  display: flex !important;
-  height: 40px;
-  justify-content: center;
-  margin-right: 10px;
-  width: 40px; }
 
 /* ====================== Border Helpers ====================== */
 ._border.top {
@@ -417,3 +309,109 @@ textarea {
   @media (min-width: 1000px) {
     textarea {
       width: 50%; } }
+
+#menu-open {
+  position: fixed;
+  left: 0;
+  bottom: 0;
+  margin: 5px; }
+
+#menu-close {
+  position: fixed;
+  left: 0;
+  bottom: 0;
+  margin: 5px;
+  z-index: 999; }
+
+.button-drawer {
+  background: rgba(0, 0, 0, 0.6);
+  bottom: -50px;
+  display: flex;
+  flex-direction: row;
+  height: 45px;
+  justify-content: space-around;
+  left: 0;
+  margin: 0 auto;
+  position: fixed;
+  width: 100%;
+  transition-property: all;
+  transition-duration: 0.75s;
+  transition-timing-function: cubic-bezier(0, 1, 0.5, 1); }
+  .button-drawer.open {
+    bottom: 0; }
+
+.button-drawer-style {
+  color: #e7e7e7;
+  font-size: 12px;
+  font-weight: 300;
+  letter-spacing: 2px;
+  text-transform: uppercase;
+  transition-duration: 0.75s;
+  transition-property: all;
+  border-bottom: 2px solid transparent;
+  align-self: center; }
+  .button-drawer-style.on {
+    color: #fff;
+    text-shadow: -1px 1px 3px rgba(208, 208, 208, 0.2), 1px -1px 3px rgba(252, 252, 252, 0.2);
+    border-bottom: 2px solid white; }
+  .button-drawer-style.dim {
+    color: rgba(255, 255, 255, 0.5); }
+
+.asset-drawer {
+  background: rgba(0, 0, 0, 0.6);
+  display: flex;
+  flex-direction: row;
+  justify-content: center;
+  align-items: center;
+  width: 100%;
+  position: fixed;
+  bottom: 45px;
+  height: 40px;
+  padding: 5px;
+  transition-property: all;
+  transition-duration: 0.75s;
+  transition-timing-function: cubic-bezier(0, 1, 0.5, 1); }
+
+.asset-row {
+  display: flex;
+  flex-direction: row; }
+
+#audio-menu-open {
+  position: fixed;
+  right: 0;
+  bottom: 0;
+  margin: 5px;
+  z-index: 999; }
+
+#audio-menu {
+  position: fixed;
+  right: 0;
+  bottom: 0;
+  width: 33%;
+  background: rgba(0, 0, 0, 0.65); }
+
+button {
+  background: none;
+  border: 0;
+  font: inherit;
+  line-height: normal;
+  outline: none;
+  overflow: visible; }
+
+.btn.gourd {
+  background: #d9d9d9;
+  border-bottom: 2px solid transparent;
+  height: 25px;
+  margin: 5px;
+  margin: 0 20px;
+  padding-top: 5px;
+  width: 25px;
+  border-radius: 50%;
+  font-size: 10px; }
+  .btn.gourd:hover {
+    background: #c8c8c8; }
+  .btn.gourd:active {
+    background: #b4b4b4; }
+  .btn.gourd.on {
+    background: #dbdbdb;
+    box-shadow: 0px 0px 5px #ebebeb; }
