--- public/assets/js/app.js
@@ -1,7 +1,0 @@
-const setup = require('./setup')
-
-const init = function () {
-  setup()
-}
-
-init()
