+++ public/assets/js/dom-elements.js
@@ -37,7 +37,10 @@ const el = {
 
   get allMenuButtons () {
     return [this.openMuzak, this.openWallpaper, this.openTypeface, this.openFieldRecording]
-  }
+  },
+
+  openAudioMenu: document.getElementById('audio-menu-open'),
+  audioMenu: document.getElementById('audio-menu')
 
 }
